Better DMlibMesh infrastructure: eliminating petsc_dm_nonlinear_solver and the need for --use-petsc-dm.
