+++ src/base/libmesh.C
@@ -63,6 +63,9 @@
 EXTERN_C_FOR_PETSC_BEGIN
 # include <petsc.h>
 # include <petscerror.h>
+#if !PETSC_RELEASE_LESS_THAN(3,3,0)
+#include "libmesh/petscdmlibmesh.h"
+#endif
 EXTERN_C_FOR_PETSC_END
 # if defined(LIBMESH_HAVE_SLEPC)
 #  include "libmesh/slepc_macro.h"
@@ -525,6 +528,15 @@ LibMeshInit::LibMeshInit (int argc, const char* const* argv,
           CHKERRABORT(libMesh::GLOBAL_COMM_WORLD,ierr);
         }
 # endif
+#if !PETSC_RELEASE_LESS_THAN(3,3,0)
+      // Register the reference implementation of DMlibMesh
+#if PETSC_RELEASE_LESS_THAN(3,4,0)
+  ierr = DMRegister(DMLIBMESH, PETSC_NULL, "DMCreate_libMesh", DMCreate_libMesh); CHKERRABORT(libMesh::GLOBAL_COMM_WORLD,ierr);
+#else
+  ierr = DMRegister(DMLIBMESH, DMCreate_libMesh); CHKERRABORT(libMesh::GLOBAL_COMM_WORLD,ierr);
+#endif
+
+#endif
     }
 #endif
